+++ prediction_market_agent_tooling/markets/omen/data_models.py
@@ -1,0 +1,66 @@
+import typing as t
+
+from eth_typing import ChecksumAddress, HexAddress
+from pydantic import BaseModel
+from web3 import Web3
+
+from prediction_market_agent_tooling.gtypes import (
+    USD,
+    OmenOutcomeToken,
+    Probability,
+    Wei,
+    xDai,
+)
+
+
+class OmenMarket(BaseModel):
+    """
+    https://aiomen.eth.limo
+    """
+
+    id: HexAddress
+    title: str
+    collateralVolume: Wei
+    usdVolume: USD
+    collateralToken: HexAddress
+    outcomes: list[str]
+    outcomeTokenAmounts: list[OmenOutcomeToken]
+    outcomeTokenMarginalPrices: t.Optional[list[xDai]]
+    fee: t.Optional[Wei]
+
+    @property
+    def market_maker_contract_address(self) -> HexAddress:
+        return self.id
+
+    @property
+    def market_maker_contract_address_checksummed(self) -> ChecksumAddress:
+        return Web3.to_checksum_address(self.market_maker_contract_address)
+
+    @property
+    def collateral_token_contract_address(self) -> HexAddress:
+        return self.collateralToken
+
+    @property
+    def collateral_token_contract_address_checksummed(self) -> ChecksumAddress:
+        return Web3.to_checksum_address(self.collateral_token_contract_address)
+
+    @property
+    def outcomeTokenProbabilities(self) -> t.Optional[list[Probability]]:
+        return (
+            [Probability(float(x)) for x in self.outcomeTokenMarginalPrices]
+            if self.outcomeTokenMarginalPrices is not None
+            else None
+        )
+
+    def get_outcome_str(self, outcome_index: int) -> str:
+        n_outcomes = len(self.outcomes)
+        if outcome_index >= n_outcomes:
+            raise ValueError(
+                f"Outcome index `{outcome_index}` not valid. There are only "
+                f"`{n_outcomes}` outcomes."
+            )
+        else:
+            return self.outcomes[outcome_index]
+
+    def __repr__(self) -> str:
+        return f"Omen's market: {self.title}"
