+++ tests/deploy/test_deploy.py
@@ -1,7 +1,7 @@
 import random
 
 from prediction_market_agent_tooling.deploy.agent import DeployableAgent
-from prediction_market_agent_tooling.markets.data_models import AgentMarket
+from prediction_market_agent_tooling.markets.agent_market import AgentMarket
 from prediction_market_agent_tooling.markets.markets import MarketType
